Extract average measurement creation to a separate module
It's complex enough that it's a good idea to separate it both from the
submission process and the tasks module.
